Switch to using primarily the relative turning encoder
Create a correspondence between the absolute angle encoder and the
turning motor's relative encoder so that it's possible to use either
one. This is a preparatory step for controlling the turning motor
*position* via the motor controller's PID controller, rather than
controlling the turning motor *velocity* via an external PID controller.

Implement/use ValueCache for swerve module encoders. This reduces CAN
bus load and reduces latency.