Use odometry rather than gyro for field-relative driving
The gyro is initialized to precisely represent the initial pose, which
tends to be good enough for teleop, assuming the robot was placed at a
reasonably close rotation. However, autonomous routines were
consistently off if the robot wasn't placed at precisely the correct
rotation. This change uses the estimated rotation, which rapidly
converges on the correct rotation if AprilTags are in view.